+++ src/core/server/effects/loadEffects.ts
@@ -1,0 +1,2 @@
+import './food';
+import './water';