+++ src/core/server/extensions/playerFunctions/sync.ts
@@ -6,6 +6,7 @@ import { SYSTEM_EVENTS } from '../../../shared/enums/system';
 import emit from './emit';
 import save from './save';
 import { DEFAULT_CONFIG } from '../../athena/main';
+import { playerFuncs } from '../Player';
 
 /**
  * Synchronize currency data like bank, cash, etc.
@@ -90,18 +91,7 @@ function food(p: alt.Player): void {
         return;
     }
 
-    if (p.data.food === undefined || p.data.food === null) {
-        p.data.food = 100;
-    }
-
-    p.data.food -= DEFAULT_CONFIG.FOOD_REMOVAL_RATE;
-
-    if (p.data.food <= 0) {
-        p.data.food = 0;
-    }
-
-    emit.meta(p, 'food', p.data.food);
-    save.field(p, 'food', p.data.food);
+    playerFuncs.safe.addFood(p, -DEFAULT_CONFIG.FOOD_REMOVAL_RATE);
 }
 
 function water(p: alt.Player): void {
@@ -111,18 +101,7 @@ function water(p: alt.Player): void {
         return;
     }
 
-    if (p.data.water === undefined || p.data.water === null) {
-        p.data.water = 100;
-    }
-
-    p.data.water -= DEFAULT_CONFIG.WATER_REMOVAL_RATE;
-
-    if (p.data.water <= 0) {
-        p.data.water = 0;
-    }
-
-    emit.meta(p, 'water', p.data.water);
-    save.field(p, 'water', p.data.water);
+    playerFuncs.safe.addWater(p, -DEFAULT_CONFIG.FOOD_REMOVAL_RATE);
 }
 
 function vehicles(p: alt.Player): void {